Bugfix - remove cast to int32 from offsets in WrappedMultivalueDataset.

+++ src/torchfm/dataset/wrapper_multivalued_dataset.py
@@ -31,7 +31,7 @@ class WrapperMultivaluedDataset(torch.utils.data.Dataset):
         global_offsets = list(range(num_columns - 1))
 
         self.weights = np.vstack(data[miltival_col_name].apply(lambda x: WrapperMultivaluedDataset._create_weights(x, self.max_length, num_columns - 1)).to_numpy())
-        self.offsets = np.vstack(data[miltival_col_name].apply(lambda x: WrapperMultivaluedDataset._create_offsets(x, num_columns - 2, global_offsets)).to_numpy().astype(np.int32))
+        self.offsets = np.vstack(data[miltival_col_name].apply(lambda x: WrapperMultivaluedDataset._create_offsets(x, num_columns - 2, global_offsets)).to_numpy())
 
         data[miltival_col_name] = data[miltival_col_name].apply(lambda x: x + [0] * (self.max_length - len(x)))
